Re-Organization for Simplification
Building the Target Directory Structure for migratation and re-writes for CY2024

Wired Automation Labs will be re-written and re-annotated to align to new Catalyst Center UI and consolidated for ease of Navigation.

Wireless Automation Lab will be relocated and updated with Catalyst Center Nomencalature and imagery as available.

Other labs will be re-aligned for ease of navigation.

ALL Code examples will be relocated to CODE folder so examples can be found in one location. Labs will use this central resource moving forward for code.

This process will extend through CY2024